# PERF: ArrowExtensionArray.searchsorted